todo_api_daniel/requirements.txt
Django==4.2
djangorestframework==3.14.0
gunicorn==23.0.0
psycopg2-binary==2.9.6
